I have implemented my previous changes into the proper scene
