+++ Game/Assets/Scripts/GameManager.cs
@@ -6,4 +6,6 @@ public class GameManager
 {
     public static int numberOfPlayers = 0;
     public static bool cookBookActive = true;
+    public static bool isTouchingBook;
+    public static int[] isStepCompleted;
 }
