+++ Game/Assets/Scripts/RecipeBookTouching.cs
@@ -1,0 +1,36 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class RecipeBookTouching : MonoBehaviour
+{
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+
+    void OnCollisionEnter(Collision collision)
+    {
+        if (collision.gameObject.tag == "CookBook")
+        {
+            Debug.Log("Touching cook book");
+            GameManager.isTouchingBook = true;
+        }
+    }
+
+    void OnCollisionExit(Collision collision)
+    {
+        if (collision.gameObject.tag == "CookBook")
+        {
+            Debug.Log("Not touching cook book");
+            GameManager.isTouchingBook = false;
+        }
+    }
+}
