+++ Game/Assets/Scripts/RecipeBookTouching.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 702f8fe063e46e94abbb37d412c4a7b6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
