+++ Game/Logs/shadercompiler-UnityShaderCompiler.exe0.log
@@ -1,4 +1,4 @@
-Base path: 'C:/Program Files/Unity/Hub/Editor/2020.3.38f1/Editor/Data', plugins path 'C:/Program Files/Unity/Hub/Editor/2020.3.38f1/Editor/Data/PlaybackEngines'
+Base path: 'C:/Program Files/Unity 2020.3.38f1/Editor/Data', plugins path 'C:/Program Files/Unity 2020.3.38f1/Editor/Data/PlaybackEngines'
 Cmd: initializeCompiler
 
 Cmd: shutdown
